renaming the folder to orgName
