+++ package.json
@@ -22,58 +22,51 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@material-ui/core": "^4.11.3",
-    "@material-ui/data-grid": "^4.0.0-alpha.17",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.53",
-    "@material-ui/styles": "^4.9.14",
-    "@types/node": "^14.0.14",
+    "@material-ui/data-grid": "^4.0.0-alpha.24",
+    "@material-ui/icons": "^4.11.2",
+    "@material-ui/lab": "^4.0.0-alpha.57",
+    "@material-ui/styles": "^4.11.3",
+    "@types/node": "^14.14.37",
     "@zeit/next-css": "^1.0.1",
     "axios": "^0.21.1",
-    "classnames": "^2.2.6",
+    "classnames": "^2.3.1",
     "compression": "^1.7.4",
     "data-grid": "^3.0.0",
-    "echarts": "^4.8.0",
-    "echarts-for-react": "^2.0.16",
-    "eslint": "^7.15.0",
+    "eslint": "^7.24.0",
     "express": "^4.17.1",
-    "immer": "^4.0.2",
-    "isomorphic-unfetch": "^3.0.0",
+    "immer": "^9.0.1",
+    "isomorphic-unfetch": "^3.1.0",
     "material-ui": "^0.20.2",
     "material-ui-dropzone": "^3.5.0",
     "material-ui-image": "^3.3.1",
-    "netlify": "^6.1.1",
-    "netlify-cli": "^3.4.2",
-    "next": "^9.5.0",
-    "next-redux-wrapper": "^4.0.1",
+    "next": "^10.1.3",
+    "next-redux-wrapper": "^6.0.2",
     "node-fetch": "^2.6.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.13.1",
-    "react-hook-form": "^6.15.4",
-    "react-player": "^2.3.1",
-    "react-redux": "^7.1.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-hook-form": "^7.0.5",
+    "react-player": "^2.9.0",
+    "react-redux": "^7.2.3",
     "react-table-filter": "^2.0.2",
-    "react-tabs": "^3.1.1",
-    "react-useanimations": "^1.2.14",
-    "recharts": "^1.8.5",
-    "redux": "^4.0.4",
+    "react-tabs": "^3.2.2",
+    "redux": "^4.0.5",
     "redux-saga": "^1.1.3",
-    "socket.io-client": "^2.3.0",
-    "swr": "^0.2.3",
+    "socket.io-client": "^4.0.1",
+    "swr": "^0.5.5",
     "typescript-fsa": "^3.0.0",
-    "typescript-fsa-reducers": "^1.2.1"
+    "typescript-fsa-reducers": "^1.2.2"
   },
   "devDependencies": {
-    "@types/next-redux-wrapper": "^2.0.2",
-    "@types/react": "^16.9.35",
-    "@types/react-redux": "^7.1.8",
+    "@types/next-redux-wrapper": "^3.0.0",
+    "@types/react": "^17.0.3",
+    "@types/react-redux": "^7.1.16",
     "@types/styled-jsx": "^2.2.8",
-    "prettier": "^1.19.1",
-    "redux-devtools-extension": "^2.13.8",
-    "serverless-next.js": "^1.11.5",
-    "tslint": "^5.20.1",
+    "prettier": "^2.2.1",
+    "redux-devtools-extension": "^2.13.9",
+    "serverless-next.js": "^1.14.0",
     "tslint-config-prettier": "^1.18.0",
-    "tslint-config-standard": "^8.0.1",
+    "tslint-config-standard": "^9.0.0",
     "tslint-plugin-prettier": "^2.3.0",
-    "typescript": "^3.9.2"
+    "typescript": "^4.2.4"
   }
 }
