+++ pages/[organizationName]/about.tsx
@@ -1,8 +1,9 @@
+
+
 import {
     Container,
     Typography,
     NoSsr,
-    Link,
   } from "@material-ui/core"
   import { makeStyles } from "@material-ui/core/styles"
   import React from "react"
@@ -30,32 +31,35 @@ import { Organization } from "."
     },
   }));
   
-  function Help({ organization }: InferGetServerSidePropsType<typeof getServerSideProps>) {
+  function About({ organization }: InferGetServerSidePropsType<typeof getServerSideProps>) {
     const classes = useStyles(organization)
 
     return (
         <NoSsr>
           {organization ? (
             <BasePage className={classes.rootLight} title={organization.name} orgId={organization.id}>
-              <title>{organization.name} | How to help</title>
+              <title>{organization.name} | About us</title>
               <div className={classes.content}>
                 <Container>
-                  <Typography  variant="h3" align="center" color="primary" gutterBottom className={classes.infoText}> Thank you for your desire to help. Changing the world is easy with the team effort!
-                  </Typography>
-                  <Typography variant="h4" align="center" color="textSecondary" style={{marginTop:"200px", marginBottom:"50px"}}>
-                    Ways you can support us:
+                  <Typography  variant="h3" align="center" color="textPrimary" gutterBottom className={classes.infoText}> About Us
                   </Typography>
                   <Typography variant="h5" align="center" color="textSecondary" paragraph>
-                   -  <Link href={`/${organization.id}/payment`}>Donate to the organization or choose a specific child to support</Link> 
+                  {organization.tagline}
                   </Typography>
-                  <Typography variant="h5" align="center" color="textSecondary" paragraph >
-                   - Subscribe to our newsletter on the homepage
+                  <Typography  variant="h3" align="center" color="textPrimary" gutterBottom className={classes.infoText}> Our Mission
                   </Typography>
-                  <Typography variant="h5" align="center" color="textSecondary" paragraph >
-                  - Share our details on your social media page
+                  <Typography variant="h5" align="center" color="textSecondary" paragraph>
+                    {organization.mission}
                   </Typography>
-                  <Typography variant="h5" align="center" color="textSecondary" paragraph >
-                  -  Volunteer your time, we have lots to enhance in our organization and need enthusiasts!
+                  <Typography  variant="h3" align="center" color="textPrimary" gutterBottom className={classes.infoText}> Our Values
+                  </Typography>
+                  <Typography variant="h5" align="center" color="textSecondary" paragraph>
+                    {organization.vision}
+                  </Typography>
+                <Typography  variant="h3" align="center" color="textPrimary" gutterBottom className={classes.infoText}> Our Story
+                  </Typography>
+                  <Typography variant="h5" align="center" color="textSecondary" paragraph>
+                  {organization.story}
                   </Typography>
                 </Container>
               </div>
@@ -71,9 +75,9 @@ import { Organization } from "."
   }
   
   export const getServerSideProps: GetServerSideProps = async context => {
-    const { organizationId} = context.query
+    const {organizationName} = context.query
   
-    const orgReq = await fetch(`${process.env.mongenCoreInternal}/api/v1/organization/${organizationId}/`, {
+    const orgReq = await fetch(`${process.env.mongenCoreInternal}/api/v1/organization/${organizationName}/`, {
       method: "GET",
     })
   
@@ -84,5 +88,5 @@ import { Organization } from "."
     }
   }
   
-  export default Help
+  export default About
