+++ pages/[organizationName]/admin.tsx
@@ -131,9 +131,9 @@ import { Organization } from "."
   }
 
   export const getServerSideProps: GetServerSideProps = async context => {
-    const { organizationId} = context.query
+    const { organizationName} = context.query
   
-    const orgReq = await fetch(`${process.env.mongenCoreInternal}/api/v1/organization/${organizationId}/`, {
+    const orgReq = await fetch(`${process.env.mongenCoreInternal}/api/v1/organization/${organizationName}/`, {
       method: "GET",
     })
