+++ pages/[organizationName]/child.tsx
