+++ pages/[organizationName]/index.tsx
@@ -198,9 +198,9 @@ function Index({ organization }: InferGetServerSidePropsType<typeof getServerSid
 }
 
 export const getServerSideProps: GetServerSideProps = async context => {
-  const { organizationId} = context.query
+  const { organizationName} = context.query
 
-  const orgReq = await fetch(`${process.env.mongenCoreInternal}/api/v1/organization/${organizationId}/`, {
+  const orgReq = await fetch(`${process.env.mongenCoreInternal}/api/v1/organization/${organizationName}/`, {
     method: "GET",
   })
