+++ pages/[organizationName]/profile.tsx
@@ -280,9 +280,9 @@ import FavoriteIcon from '@material-ui/icons/Favorite'
   }
   
   export const getServerSideProps: GetServerSideProps = async context => {
-    const { organizationId} = context.query
+    const { organizationName} = context.query
   
-    const orgReq = await fetch(`${process.env.mongenCoreInternal}/api/v1/organization/${organizationId}/`, {
+    const orgReq = await fetch(`${process.env.mongenCoreInternal}/api/v1/organization/${organizationName}/`, {
       method: "GET",
     })
